Add tests for Process.WaitForExit
Our Unix implementation of Process.WaitForExit is relatively complicated, as we need to support functionality implicitly supported by Windows that's not built-in on Unix.  This commit adds some tests to more heavily exercise Process waiting; I plan to add some more subsequently, but I wanted to get these in now.