Disable parallelization of FSW tests
My previous change to rename FSW test classes led to FSW tests getting parallelized.  It turns out that some of the tests are not robust against this due to using file names that can end up conflicting with other tests.  Additionally, it appears that the waiting / timeouts used by the tests is very susceptible to failure when run in parallel due to reliance on the shared thread pool, causing more failures in CI.