HttpClient xplat: Fix multiple issues with socket/timer callbacks
In earlier version, a curl_multi_socket_action was performed everytime
the callback was invoked irrespective of whether there was activity on
the socket. This in turn causes more callbacks when libcurl tries to read
the socket but finds no data. This can be made more efficient by waiting
for activity on the socket and informing libcurl accordingly
- Fixed implementation of timer callback using an actual timer and handling
  all possible values of timeout
- Fixed a bug in CheckForCompletedTransfers where a mismatch could occur
  between the easy handle specified in the DONE msg
- Fixed the SEGV which was happening in earlier code due to recursive calls
  to curl_multi_socket_action in the socket callback which was being invoked
  synchronously in curl_multi_socket_action. This was causing a stack overflow
  leading to a segfault