Fix bug in Semaphore.NamedProducerConsumer test
The test schedules two tasks, each of which creates a Semaphore with the same name.  One of the tasks releases count into the semaphore, and the other consumes it via WaitOne.  If the first task ends up creating the semaphore first, there's no problem: the first task will end up blocking, and the second task will end up creating the same semaphore and releasing count into it.  But if the second task ends up executing first, it could create the semaphore, release count into it, and destroy the semaphore, all before the first task constructs the semaphore.  If that happens, the first task will construct an entirely different semaphore, and end up blocking until timeout, at which point the test will fail.

This fix introduces a barrier that ensures neither task waits or releases until both tasks have constructed the shared semaphore.