Add ProcessStartInfo.PasswordInClearText to System.Diagnostics.Process portable library
System.Diagnostics.Process took dependency on System.Security.SecureString for credentials support. In the recent API Review (https://github.com/dotnet/apireviews/pull/11) we decided to remove the dependency by adding a new property to ProcessStartInfo.PasswordInClearText that has a return type of string.

This is achieved in two phases.
     1. Remove credentials from System.Diagnostics.Process v4.0.0.0 contract to ensure that the contract can be supported on Net4.6
     2. Add ProcessStartInfo.PasswordInClearText property and not expose any static Process.Start APIs for this scenario (these are advances sceanarios anyway).

[tfs-changeset: 1516848]