Implement cancelable ReadAsync operations
With the infrastructure changes out of the way, we can now add more robust cancellation to ReadAsync operations.

Today when a CancellationToken is passed to ReadAsync, we do an initial check for whether cancellation has been requested, and then we call read, which will block until data is available, in a non-cancelable fashion.

This change makes that read cancelable.  If a cancelable token is provided, we create an anonymous pipe for the operation.  One end of the pipe is written to when cancellation is requested, and rather than doing a blocking read operation, we instead do a blocking poll operation that includes both the underlying file descriptor as well as the read end of this special one-time pipe.  If the poll wakes up because there's data available to read, we read.  If it wakes up because the cancellation pipe was signaled, we throw an OperationCanceledException.

We unfortunately can't enable WriteAsync cancellation in the same way, but luckily WriteAsync is less likely to block.