Use AuthorityInformationAccess to download intermediate CA certificates on Unix.
When the issuer certificate cannot be found in the locally understood stores, use the AIA record to extract the URL for the issuer certificate.  Then use libcurl to download it, and apply it into the chain candidates set.  If the chain then is considered successful save those certificates that had to be downloaded to the "CA" (intermediate) store for the current user.

The timeout value can be exceeded, but once it has been exceeded no further downloads will take place.  The same timeout will be used in the future for CRL/OCSP checks.

No unit tests were added for this feature because they would require both the network (downloading the files) and the filesystem (deleting a file to force it to download, then the product code causing it to be written).