Fully import Contains, DefaultIfEmpty, Distinct and SequenceEquals tests
Convert Contains, DefaultIfEmpty, Distinct and SequenceEquals tests imported
in #2874 into a more directly XUnit-ian style.

Renamed tests (and file) for SequnceEqual from EqualAllTests to
SequenceEqualTests

Comment on Contains2e() suggests it intended to use a range repeat than a range
as it did, so altered accordingly in PresentMultipleTimesNotICollectionT(). The
case it had been covering is already covered by
FirstElementMatchesNotICollectionT()

Added throw-on-null tests for DefaultIfEmpty, Contains and SequenceEqual