Avoid blocking CurlHandler in read callback
PR #3105 addressed the issue we were facing with libcurl and the write callback used when libcurl has response body data available.  We were blocking libcurl's callback until a reader was available to consume the data, resulting in both significant scalability issues as well as the potential for easy deadlock.  This was fixed by making the whole operation async via libcurl's ability to pause connections.

The same issue exists for libcurl's read callback, used by libcurl to read data from the request stream to send as part of the request.  The issue here isn't as impactful as on the write side, because a) a common case has no request stream or has a MemoryStream that serves up its data immediately, and b) other streams (e.g. FileStream) typically serve their data quickly even if blocking is involved and don't interact with the HttpClient in a way that could lead to deadlock.  However, it's still an issue, and it's still possible to construct cases where deadlock is possible, e.g. if the output stream of one request is used as the input stream of another request associated with the same libcurl multi handle.

This commit addresses the read callback case by relying on the same pause mechanism used for the write callback.  When libcurl requests data, if the source is a memory stream, we just do the read and hand back the data. Otherwise, we issue a ReadAsync on the stream.  If the operation completes immediately, we synchronously hand the data back to libcurl, storing any extra beyond what libcurl asked for.  If the operation didn't complete immediately, we store it for later, and pause the connection, requesting an unpause when the read operation completes.  Subsequent read requests from libcurl first try to be satisfied from any previous ReadAsync with
data remaining.

As part of this, I also removed the pause state tracking I'd previously added.  In my previous PR as part of adding support for pausing, I added a field to EasyRequest to attempt to mirror the current paused state as known by libcurl.  This was done to try to avoid requesting unpausing in cases where we'd like to be unpaused but have already requested it. However, this appears to have been premature optimization, as the case is relatively rare, the overheads relatively low for processing the extra request, and it causes additional headaches once we're able to pause sending/receiving individually.